requirements.txt
djangorestframework==3.8.2
pytz==2018.5
sqlparse==0.2.4
